re-add resources so we can track them
In a previous commit, it was assumed that the best way to
incorporate external libraries and resources was to list them in
package.json, `npm install` them, and then reference those files
directly from the add-on. Unfortunately, AMO reviewers didn't like
this, because doing so would bring in many unused and unrelated
files (I guess they want to make more work for themselves by
reviewing these additional libraries rather than just checking that
they were incorporated as-is and so are as trustworthy as when
anyone runs `npm install`). ¯\_(ツ)_/¯